Relocating automower clients.
