+++ src/automowerAccessory.ts
@@ -2,8 +2,8 @@ import {
     PlatformAccessory, UnknownContext
 } from 'homebridge';
 
-import { Mower } from './clients/automowerClient';
-import { SettingsEvent, StatusEvent } from './clients/automowerEventStreamClient';
+import { Mower } from './clients/automower/automowerClient';
+import { SettingsEvent, StatusEvent } from './clients/automower/automowerEventStreamClient';
 import { AccessoryInformation } from './services/accessoryInformation';
 import { ArrivingSensor } from './services/arrivingSensor';
 import { BatteryInformation } from './services/batteryInformation';
