+++ src/automowerAccessoryFactory.ts
@@ -1,7 +1,7 @@
 import { API, PlatformAccessory } from 'homebridge';
 
 import { AutomowerAccessory, AutomowerContext } from './automowerAccessory';
-import { Mower } from './clients/automowerClient';
+import { Mower } from './clients/automower/automowerClient';
 import { PlatformLogger } from './diagnostics/platformLogger';
 import { Localization } from './primitives/localization';
 import { PlatformAccessoryFactory } from './primitives/platformAccessoryFactory';
