+++ src/automowerPlatform.ts
@@ -5,7 +5,7 @@ import {
 
 import { AutomowerAccessory, AutomowerContext } from './automowerAccessory';
 import { AutomowerAccessoryFactory, AutomowerAccessoryFactoryImpl } from './automowerAccessoryFactory';
-import { SettingsEvent, StatusEvent } from './clients/automowerEventStreamClient';
+import { SettingsEvent, StatusEvent } from './clients/automower/automowerEventStreamClient';
 import { BadConfigurationError } from './errors/badConfigurationError';
 import { AuthenticationMode, DeviceType } from './model';
 import { Localization, Y18nLocalization } from './primitives/localization';
