+++ src/clients/automower/automowerEventStreamClient.ts
@@ -1,8 +1,8 @@
-import { PlatformLogger } from '../diagnostics/platformLogger';
-import { AccessToken } from '../model';
-import { PLUGIN_ID } from '../settings';
+import { PlatformLogger } from '../../diagnostics/platformLogger';
+import { AccessToken } from '../../model';
+import { PLUGIN_ID } from '../../settings';
+import { WebSocketWrapper, WebSocketWrapperImpl } from '../primitives/webSocketWrapper';
 import { Battery, Calendar, Headlight, MowerMetadata, MowerState, Planner, Position } from './automowerClient';
-import { WebSocketWrapper, WebSocketWrapperImpl } from './primitives/webSocketWrapper';
 
 /**
  * Describes a connected event.
