+++ src/primitives/platformContainer.ts
@@ -8,10 +8,10 @@ import * as settings from '../settings';
 import { AutomowerAccessoryFactoryImpl } from '../automowerAccessoryFactory';
 import { AutomowerPlatformConfig } from '../automowerPlatform';
 import { AuthenticationClientImpl } from '../clients/authenticationClient';
-import { AutomowerClientImpl } from '../clients/automowerClient';
-import { AutomowerEventStreamClientImpl } from '../clients/automowerEventStreamClient';
+import { AutomowerClientImpl } from '../clients/automower/automowerClient';
+import { AutomowerEventStreamClientImpl } from '../clients/automower/automowerEventStreamClient';
+import { RateLimitedAutomowerClient } from '../clients/automower/rateLimitedAutomowerClient';
 import { RetryerFetchClient } from '../clients/fetchClient';
-import { RateLimitedAutomowerClient } from '../clients/rateLimitedAutomowerClient';
 import { HomebridgeImitationLogger } from '../diagnostics/platformLogger';
 import { ConsoleWrapperImpl } from '../diagnostics/primitives/consoleWrapper';
 import { DefaultErrorFactory } from '../errors/errorFactory';
