+++ src/services/homebridge/abstractContactSensor.ts
@@ -1,7 +1,7 @@
 import { API, Characteristic, PlatformAccessory, Service } from 'homebridge';
 
 import { AutomowerContext } from '../../automowerAccessory';
-import { MowerMetadata } from '../../clients/automowerClient';
+import { MowerMetadata } from '../../clients/automower/automowerClient';
 import { PlatformLogger } from '../../diagnostics/platformLogger';
 import { Policy } from '../policies/policy';
 import { AbstractAccessoryService } from './abstractAccessoryService';
