+++ src/services/husqvarna/automower/getMowersService.ts
@@ -1,4 +1,4 @@
-import { AutomowerClient, Mower } from '../../../clients/automowerClient';
+import { AutomowerClient, Mower } from '../../../clients/automower/automowerClient';
 import { NotAuthorizedError } from '../../../errors/notAuthorizedError';
 import { AccessTokenManager } from '../accessTokenManager';
