+++ src/services/motionSensor.ts
@@ -1,7 +1,7 @@
 import { API, Characteristic, PlatformAccessory, Service } from 'homebridge';
 
 import { AutomowerContext } from '../automowerAccessory';
-import { MowerMetadata, MowerState } from '../clients/automowerClient';
+import { MowerMetadata, MowerState } from '../clients/automower/automowerClient';
 import { PlatformLogger } from '../diagnostics/platformLogger';
 import { AbstractAccessoryService } from './homebridge/abstractAccessoryService';
 import { MowerFaultedPolicy } from './policies/mowerFaultedPolicy';
