+++ src/services/pauseSwitch.ts
@@ -1,6 +1,6 @@
 import { API, CharacteristicSetCallback, HAPStatus, PlatformAccessory } from 'homebridge';
 import { AutomowerContext } from '../automowerAccessory';
-import { Activity, MowerState, State } from '../clients/automowerClient';
+import { Activity, MowerState, State } from '../clients/automower/automowerClient';
 import { PlatformLogger } from '../diagnostics/platformLogger';
 import { AbstractSwitch, Switch } from './homebridge/abstractSwitch';
 import { MowerControlService } from './husqvarna/automower/mowerControlService';
