+++ src/services/scheduleSwitch.ts
@@ -4,7 +4,7 @@ import {
 } from 'homebridge';
 
 import { AutomowerContext } from '../automowerAccessory';
-import { Calendar, MowerState, Planner } from '../clients/automowerClient';
+import { Calendar, MowerState, Planner } from '../clients/automower/automowerClient';
 import { PlatformLogger } from '../diagnostics/platformLogger';
 import { AbstractSwitch, Switch } from './homebridge/abstractSwitch';
 import { attachCuttingHeightCharacteristic } from './homebridge/characteristics/cuttingHeight';
