+++ test/services/homebridge/abstractSwitch.spec.ts
@@ -3,7 +3,7 @@ import { API, PlatformAccessory } from 'homebridge';
 import { It, Mock, Times } from 'moq.ts';
 
 import { AutomowerContext } from '../../../src/automowerAccessory';
-import { MowerMetadata } from '../../../src/clients/automowerClient';
+import { MowerMetadata } from '../../../src/clients/automower/automowerClient';
 import { PlatformLogger } from '../../../src/diagnostics/platformLogger';
 import { NameMode } from '../../../src/services/homebridge/abstractSwitch';
 import { SwitchSpy } from './switchSpy';
