+++ test/services/husqvarna/automower/mowerControlService.spec.ts
@@ -1,6 +1,6 @@
 import { It, Mock, Times } from 'moq.ts';
 
-import { AutomowerClient } from '../../../../src/clients/automowerClient';
+import { AutomowerClient } from '../../../../src/clients/automower/automowerClient';
 import { NotAuthorizedError } from '../../../../src/errors/notAuthorizedError';
 import { AccessToken } from '../../../../src/model';
 import { AccessTokenManager } from '../../../../src/services/husqvarna/accessTokenManager';
