+++ test/services/husqvarna/discoveryService.spec.ts
@@ -3,7 +3,7 @@ import { It, Mock, Times } from 'moq.ts';
 import { AutomowerAccessory } from '../../../src/automowerAccessory';
 import { AutomowerAccessoryFactory } from '../../../src/automowerAccessoryFactory';
 import { AutomowerPlatform } from '../../../src/automowerPlatform';
-import { Activity, HeadlightMode, Mode, Mower, OverrideAction, RestrictedReason, State } from '../../../src/clients/automowerClient';
+import { Activity, HeadlightMode, Mode, Mower, OverrideAction, RestrictedReason, State } from '../../../src/clients/automower/automowerClient';
 import { PlatformLogger } from '../../../src/diagnostics/platformLogger';
 import { GetMowersService } from '../../../src/services/husqvarna/automower/getMowersService';
 import { DiscoveryServiceImpl } from '../../../src/services/husqvarna/discoveryService';
